+++ lib/cocoapods/installer.rb
@@ -399,7 +399,7 @@ module Pod
       UI.message "- Writing Xcode project file to #{UI.path sandbox.project_path}" do
         pods_project.pods.remove_from_project if pods_project.pods.empty?
         pods_project.development_pods.remove_from_project if pods_project.development_pods.empty?
-        pods_project.sort
+        pods_project.sort({:groups_position => :below})
         pods_project.recreate_user_schemes(false)
         pods_project.save
       end